sc-7549 GDS USER UI: uniformize the login/register style
