+++ web/gds-user-ui/src/components/Section/CreateAccount.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import { Flex, Box, Stack, Button, Text, useColorModeValue } from '@chakra-ui/react';
+import { Box, Stack, Button, Text, useColorModeValue } from '@chakra-ui/react';
 
 import { GoogleIcon } from 'components/Icon';
-import { colors } from 'utils/theme';
 import { useForm } from 'react-hook-form';
 
 import { yupResolver } from '@hookform/resolvers/yup';
@@ -13,6 +12,7 @@ import * as yup from 'yup';
 import { Trans } from '@lingui/react';
 import { t } from '@lingui/macro';
 import ChakraRouterLink from 'components/ChakraRouterLink';
+import AuthLayout from 'layouts/AuthLayout';
 
 interface CreateAccountProps {
   handleSocialAuth: (event: React.FormEvent, type: string) => void;
@@ -47,102 +47,84 @@ const CreateAccount: React.FC<CreateAccountProps> = (props) => {
   const watchPassword = watch('password');
 
   return (
-    <Flex
-      align={'center'}
-      justify={'center'}
-      fontFamily={'open sans'}
-      fontSize={'lg'}
-      mb={'10vh'}
-      bg={useColorModeValue('white', 'gray.800')}>
-      <Stack spacing={8} mx={'auto'} maxW={'lg'} py={12} px={6}>
-        <Stack align={'center'}>
-          <Text color={useColorModeValue('gray.600', 'white')}>
-            <Text as={'span'} fontWeight={'bold'}>
-              <Trans id="Create your TRISA account.">Create your TRISA account.</Trans>
-            </Text>{' '}
-            <Trans id="We recommend that a senior compliance officer initially creates the account for the VASP. Additional accounts can be created later.">
-              We recommend that a senior compliance officer initially creates the account for the
-              VASP. Additional accounts can be created later.
-            </Trans>
+    <AuthLayout>
+      <Text color={useColorModeValue('gray.600', 'white')}>
+        <Text as={'span'} fontWeight={'bold'}>
+          <Trans id="Create your TRISA account.">Create your TRISA account.</Trans>
+        </Text>{' '}
+        <Trans id="We recommend that a senior compliance officer initially creates the account for the VASP. Additional accounts can be created later.">
+          We recommend that a senior compliance officer initially creates the account for the VASP.
+          Additional accounts can be created later.
+        </Trans>
+      </Text>
+
+      <Box>
+        <Button
+          bg={'gray.100'}
+          w="100%"
+          onClick={(event: any) => props.handleSocialAuth(event, 'google')}
+          size="lg"
+          borderRadius="none"
+          color={'gray.600'}
+          _hover={{
+            background: useColorModeValue('gray.200', 'black'),
+            color: useColorModeValue('gray.600', 'white')
+          }}>
+          <GoogleIcon h={24} />
+          <Text as={'span'} ml={3}>
+            <Trans id="Continue with Google">Continue with Google</Trans>
           </Text>
-        </Stack>
-        <Stack align={'center'} justify={'center'} fontFamily={'open sans'}>
-          <Button
-            bg={'gray.100'}
-            w="100%"
-            onClick={(event: any) => props.handleSocialAuth(event, 'google')}
-            size="lg"
-            borderRadius="none"
-            color={'gray.600'}
-            _hover={{
-              background: useColorModeValue('gray.200', 'black'),
-              color: useColorModeValue('gray.600', 'white')
-            }}>
-            <GoogleIcon h={24} />
-            <Text as={'span'} ml={3}>
-              <Trans id="Continue with Google">Continue with Google</Trans>
+        </Button>
+      </Box>
+      <Text textAlign="center">Or</Text>
+      <Box bg={useColorModeValue('white', 'transparent')}>
+        <form onSubmit={handleSubmit(props.handleSignUpWithEmail)} noValidate>
+          <Stack spacing={4}>
+            <InputFormControl
+              controlId="username"
+              {...register('username')}
+              size="lg"
+              data-testid="username-field"
+              placeholder={t`Email Address`}
+              isInvalid={!!getValueByPathname(errors, 'username')}
+              formHelperText={getValueByPathname(errors, 'username')?.message}
+            />
+
+            <InputFormControl
+              controlId="password"
+              {...register('password')}
+              data-testid="password-field"
+              placeholder={t`Password`}
+              hasBtn
+              size="lg"
+              handleFn={handleClick}
+              setBtnName={show ? 'Hide' : 'Show'}
+              isInvalid={!!getValueByPathname(errors, 'password')}
+              type={show ? 'text' : 'password'}
+              formHelperText={watchPassword ? <PasswordStrength data={watchPassword} /> : null}
+            />
+            <Button
+              display="block"
+              alignSelf="center"
+              bg={'blue'}
+              color={'white'}
+              type="submit"
+              isLoading={props.isLoading}
+              _hover={{
+                bg: '#10aaed'
+              }}>
+              <Trans id="Create an Account">Create an Account</Trans>
+            </Button>
+            <Text textAlign="center">
+              <Trans id="Already have an account?">Already have an account?</Trans>{' '}
+              <ChakraRouterLink to={'/auth/login'} color="link" _hover={{ textDecor: 'underline' }}>
+                <Trans id="Log in.">Log in.</Trans>
+              </ChakraRouterLink>
             </Text>
-          </Button>
-          <Text py={3}>Or</Text>
-        </Stack>
-
-        <Box
-          rounded={'lg'}
-          bg={useColorModeValue('white', 'transparent')}
-          width={'100%'}
-          position={'relative'}
-          bottom={5}>
-          <form onSubmit={handleSubmit(props.handleSignUpWithEmail)} noValidate>
-            <Stack spacing={4}>
-              <InputFormControl
-                controlId="username"
-                {...register('username')}
-                size="lg"
-                data-testid="username-field"
-                placeholder={t`Email Address`}
-                isInvalid={!!getValueByPathname(errors, 'username')}
-                formHelperText={getValueByPathname(errors, 'username')?.message}
-              />
-
-              <InputFormControl
-                controlId="password"
-                {...register('password')}
-                data-testid="password-field"
-                placeholder={t`Password`}
-                hasBtn
-                size="lg"
-                handleFn={handleClick}
-                setBtnName={show ? 'Hide' : 'Show'}
-                isInvalid={!!getValueByPathname(errors, 'password')}
-                type={show ? 'text' : 'password'}
-                formHelperText={watchPassword ? <PasswordStrength data={watchPassword} /> : null}
-              />
-              <Button
-                display="block"
-                alignSelf="center"
-                bg={'blue'}
-                color={'white'}
-                type="submit"
-                isLoading={props.isLoading}
-                _hover={{
-                  bg: '#10aaed'
-                }}>
-                <Trans id="Create an Account">Create an Account</Trans>
-              </Button>
-              <Text textAlign="center">
-                <Trans id="Already have an account?">Already have an account?</Trans>{' '}
-                <ChakraRouterLink
-                  to={'/auth/login'}
-                  color="link"
-                  _hover={{ textDecor: 'underline' }}>
-                  <Trans id="Log in.">Log in.</Trans>
-                </ChakraRouterLink>
-              </Text>
-            </Stack>
-          </form>
-        </Box>
-      </Stack>
-    </Flex>
+          </Stack>
+        </form>
+      </Box>
+    </AuthLayout>
   );
 };
